requirements.txt
fastapi>=0.68.0
uvicorn
sqlalchemy>=1.4.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
httpx
pydantic-settings>=2.0.0
bcrypt==4.0.1